Components/ItemsList: extract functionality
