--- src/components/Item/index.js
@@ -1,1 +1,0 @@
-export { default } from './Item';
