+++ src/components/ItemsList/ItemsList.scss
@@ -1,0 +1,6 @@
+@import "../../styles/variables";
+
+.items-list {
+    overflow: hidden;
+    margin-top: $header-height;
+}
