+++ src/components/ItemsListRow/index.js
@@ -1,0 +1,1 @@
+export { default } from './ItemsListRow';
