+++ src/pages/Item/Item.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import './Item.scss';
+
+export default function Item ({ item }) {
+    return (
+        <div className='item'>
+
+        </div>
+    );
+}
