+++ src/routes.js
@@ -10,16 +10,12 @@ import ProtectedRoute from './components/ProtectedRoute';
 
 const Routes = () => (
     <div>
-        <Route path="/create-item"><NewEntry /></Route>
+        <ProtectedRoute path="/create-item"><NewEntry /></ProtectedRoute>
         <ProtectedRoute path="/items"><Items /></ProtectedRoute>
+        <ProtectedRoute path="/item"><Items /></ProtectedRoute>
         <Route path="/unlock"><Unlock /></Route>
         <Route path="/init"><Init /></Route>
         <Route path="/"><EntryPoint /></Route>
-        {/*<Fragment forRoute="/create-item"><NewEntry /></Fragment>
-        <Fragment forRoute="/items"><Items /></Fragment>
-        <Fragment forRoute="/unlock"><Unlock /></Fragment>
-        <Fragment forRoute="/init"><Init /></Fragment>
-        <Fragment forRoute="/"><EntryPoint /></Fragment>*/}
     </div>
 )
